+++ apps/web/src/app/background-tasks/page.tsx
@@ -1,0 +1,5 @@
+import { BackgroundTasksPage } from '@/components/background-tasks/BackgroundTasksPage';
+
+export default function BackgroundTasks() {
+  return <BackgroundTasksPage />;
+}